* New: improve audio volume normalizer
New volume normalizer is so-called 'dynamic audio normalizer'
which is taken from https://github.com/lordmulder/DynamicAudioNormalizer .
Basically, it's not possible to implement a 'real' normalizer in real time
because it requires all audio data at once.
This normalizer is a 'local noramlizer' which looks certain amount of time interval
before and after current playing time.